Refactor Artist and Song files with Concerns' modules
