+++ lib/artist.rb
@@ -1,30 +1,24 @@
 require 'pry'
 
 class Artist
+
   attr_accessor :name
   attr_reader :songs
 
+  extend Memorable::ClassMethods
+  extend Findable::ClassMethods
+  include Paramable::InstanceMethods
+  include Memorable::InstanceMethods
+
   @@artists = []
 
-  def self.find_by_name(name)
-    @@artists.detect{|a| a.name == name}
-  end
-
   def initialize
-    @@artists << self
+    super
     @songs = []
   end
 
   def self.all
     @@artists
-  end
-
-  def self.reset_all
-    self.all.clear
-  end
-
-  def self.count
-    @@artists.count
   end
 
   def add_song(song)
@@ -34,10 +28,6 @@ class Artist
 
   def add_songs(songs)
     songs.each { |song| add_song(song) }
-  end
-
-  def to_param
-    name.downcase.gsub(' ', '-')
   end
 
 end
