+++ lib/concerns/findable.rb
@@ -1,0 +1,9 @@
+
+module Findable
+
+  module ClassMethods
+    def find_by_name(name)
+      self.all.detect{|a| a.name == name}
+    end
+  end
+end
