+++ lib/concerns/paramable.rb
@@ -1,0 +1,10 @@
+
+module Paramable
+
+  module InstanceMethods
+    def to_param
+      name.downcase.gsub(' ', '-')
+    end
+  end
+
+end
