+++ spec/artist_spec.rb
@@ -1,4 +1,4 @@
-require_relative './spec_helper'
+require_relative './spec_helper.rb'
 
 describe Artist do
   before(:each) do
